refactors point methods [clang]
COMMENTS:
refactors the methods of the point type to improve readability and
organization

introduces the point namespace to invoke ``static''-like (not type
bounded) methods in the OOP sense such as xcompare() and ycompare()

code passes existing tests

valgrind reports no memory issues
